feature: rename expression.ts file
